+++ access/resource_secret_scope.go
@@ -66,7 +66,7 @@ func (a SecretScopesAPI) Create(s SecretScope) error {
 			//lint:ignore ST1005 Azure is a valid capitalized string
 			return fmt.Errorf("Azure KeyVault is not available")
 		}
-		if a.client.AzureAuth.IsClientSecretSet() {
+		if a.client.IsAzureClientSecretSet() {
 			//lint:ignore ST1005 Azure is a valid capitalized string
 			return fmt.Errorf("Azure KeyVault cannot yet be configured for Service Principal authorization")
 		}
@@ -123,7 +123,7 @@ func kvDiffFunc(ctx context.Context, diff *schema.ResourceDiff, v interface{}) e
 		return nil
 	}
 	client := v.(*common.DatabricksClient)
-	if client.IsAzure() && client.AzureAuth.IsClientSecretSet() {
+	if client.IsAzure() && client.IsAzureClientSecretSet() {
 		return fmt.Errorf("you can't set up Azure KeyVault-based secret scope via Service Principal")
 	}
 	return nil